+++ src/components/EventsList/index.js
@@ -1,0 +1,11 @@
+import React, { PropTypes } from 'react'
+
+class EventsList extends React.Component {
+    render () {
+        return(
+            <h1>First event</h1>
+        );
+    }
+}
+
+export default EventsList;
