+++ src/components/Sidebar/style.scss
