+++ src/main.scss
@@ -1,5 +1,5 @@
-@import "view/Sidebar/style";
-@import "view/Dashboard/style";
+@import "components/Sidebar/style";
+@import "components/Dashboard/style";
 @import "styles/vars";
 
 body {
